ESILCO-455, ESILCO-329 fix

+++ mes-plugins/mes-plugins-material-flow-resources/src/main/java/com/qcadoo/mes/materialFlowResources/listeners/PalletStorageStateListListeners.java
@@ -71,11 +71,11 @@ public class PalletStorageStateListListeners {
 
     public void showPallestWithFreeSpace(final ViewDefinitionState view, final ComponentState state, final String[] args) {
         GridComponent palletGrid = (GridComponent) view.getComponentByReference("grid");
-        BigDecimal palletWithFreePlace = dataDefinitionService.get(MaterialFlowResourcesConstants.PLUGIN_IDENTIFIER,MaterialFlowResourcesConstants.MODEL_DOCUMENT_POSITION_PARAMETERS).find().setMaxResults(1).uniqueResult().getDecimalField(DocumentPositionParametersItemValues.PALLET_WITH_FREE_PALECE);
+        Integer palletWithFreePlace = dataDefinitionService.get(MaterialFlowResourcesConstants.PLUGIN_IDENTIFIER,MaterialFlowResourcesConstants.MODEL_DOCUMENT_POSITION_PARAMETERS).find().setMaxResults(1).uniqueResult().getIntegerField(DocumentPositionParametersItemValues.PALLET_WITH_FREE_PALECE);
         if(palletWithFreePlace == null){
             palletGrid.addMessage("materialFlowResources.pallet.missing.parameter.palletWithFreeSpace.error", ComponentState.MessageType.FAILURE);
         }else {
-            palletGrid.setCustomRestriction(searchBuilder -> searchBuilder.add(SearchRestrictions.lt(PalletStorageStateDtoFields.TOTAL_QUANTITY, palletWithFreePlace)));
+            palletGrid.setCustomRestriction(searchBuilder -> searchBuilder.add(SearchRestrictions.lt(PalletStorageStateDtoFields.TOTAL_QUANTITY,  BigDecimal.valueOf(palletWithFreePlace))));
         }
 
     }
@@ -87,11 +87,11 @@ public class PalletStorageStateListListeners {
 
     public void showPalletsWithProductToShift(final ViewDefinitionState view, final ComponentState state, final String[] args) {
         GridComponent palletGrid = (GridComponent) view.getComponentByReference("grid");
-        BigDecimal palletToShift = dataDefinitionService.get(MaterialFlowResourcesConstants.PLUGIN_IDENTIFIER,MaterialFlowResourcesConstants.MODEL_DOCUMENT_POSITION_PARAMETERS).find().setMaxResults(1).uniqueResult().getDecimalField(DocumentPositionParametersItemValues.PALLET_TO_SHIFT);
+        Integer palletToShift = dataDefinitionService.get(MaterialFlowResourcesConstants.PLUGIN_IDENTIFIER,MaterialFlowResourcesConstants.MODEL_DOCUMENT_POSITION_PARAMETERS).find().setMaxResults(1).uniqueResult().getIntegerField(DocumentPositionParametersItemValues.PALLET_TO_SHIFT);
         if(palletToShift == null){
             palletGrid.addMessage("materialFlowResources.pallet.missing.parameter.palletToShift.error", ComponentState.MessageType.FAILURE);
         }else {
-            palletGrid.setCustomRestriction(searchBuilder -> searchBuilder.add(SearchRestrictions.lt(PalletStorageStateDtoFields.TOTAL_QUANTITY, palletToShift)));
+            palletGrid.setCustomRestriction(searchBuilder -> searchBuilder.add(SearchRestrictions.lt(PalletStorageStateDtoFields.TOTAL_QUANTITY, BigDecimal.valueOf(palletToShift))));
         }
 
     }
